[Fix]#1 - remove sampleView
