--- Projects/UI/Sources/Extensions/UIImage+Extensions.swift
@@ -1,10 +1,0 @@
-//
-//  UIImage+Extensions.swift
-//  ProjectDescriptionHelpers
-//
-//  Created by 최효원 on 2023/10/06.
-//
-
-import Foundation
-
-import SwiftUI
